+++ examples/helloPlayground/helloThreadsWorker.ts
@@ -1,0 +1,7 @@
+import { expose } from 'threads/worker';
+
+expose({
+    hashPassword(password) {
+        return password + '_fake';
+    }
+});