+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## 2.0.0
+
+- Move to vite/rollup for development and building
+- Switch to TypeScript
+- Use eslint
+- Evaluate Comlink, threads.js and Objectbuffer
+
 ## 1.0.1
 - `three.js` is no longer a **peerDependency**. It is just a **dependency**.
-- Updated **devDependencies** to resolve potential security issues. 
+- Updated **devDependencies** to resolve potential security issues.
 
 ## 1.0.0
 - Initial public release.